+++ lib/once/shared.ex
@@ -9,6 +9,9 @@ defmodule Once.Shared do
   def encode16(value), do: Base.encode16(value, case: :lower)
   def decode16(value), do: Base.decode16(value, case: :mixed)
 
+  def encode32(value), do: Base.hex_encode32(value, case: :lower, padding: false)
+  def decode32(value), do: Base.hex_decode32(value, case: :mixed, padding: false)
+
   def do_to_format!(format_result, input)
   def do_to_format!({:ok, value}, _input), do: value
 
@@ -19,8 +22,10 @@ defmodule Once.Shared do
   if function_exported?(Base, :url_valid64?, 2) do
     def valid64?(bin), do: Base.url_valid64?(bin, padding: false)
     def valid16?(bin), do: Base.valid16?(bin, case: :mixed)
+    def valid32?(bin), do: Base.hex_valid32?(bin, case: :mixed, padding: false)
   else
     def valid64?(bin), do: match?({:ok, _}, decode64(bin))
     def valid16?(bin), do: match?({:ok, _}, decode16(bin))
+    def valid32?(bin), do: match?({:ok, _}, decode32(bin))
   end
 end